Validate outputDir and default to session dir + networkName + datetime
We were going to always use that default as outputDir but that would
require the user to have a saved session first. Instead we only provide
a default and let the user pick any dir they want. Because the user can
pick a dir I added validation for it too.