Add conda path setting (#12)
When given a relative path, find it on the system path, trying a couple
of extensions. When the default value, also try looking in the default
conda install locations. Otherwise, don't do anything fancy and only try
the abs path. Show an error explaining how to change the conda path
(based on CyProperty, so setting is available in cytoscape Preferences).