Fix parsing a windows path while finding conda
and use Path.of instead of deprecated Paths.get.

In particular the windows file path parser complains about encountering
':' in the path. This is because Paths.get(a, ...) expects parts of a
path not subpaths, e.g. a file name may not contain ':'. Paths.get(a)
treats it as an actual path string and thus won't complain about ':'.